+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Piral Changelog
 
+## 0.6.1 (to be determined)
+
+* Fixed externals resolution (#33)
+* Included custom metadata from pilets (#34)
+* Added ability to configure the used `history` (#37)
+
 ## 0.6.0 (July 22, 2019)
 
 * Improved the sample